refactor(OAuth2): organize oauth2 classes by directory
