+++ src/main/java/com/sequence/anonymous/security/userinfo/GithubOAuth2UserInfo.java
@@ -1,18 +1,18 @@
-package com.sequence.anonymous.security;
+package com.sequence.anonymous.security.userinfo;
 
 import java.util.Map;
 
-public class GoogleOAuth2UserInfo implements CustomOAuth2UserInfo {
+public class GithubOAuth2UserInfo implements CustomOAuth2UserInfo {
 
   private final Map<String, Object> attributes;
 
-  public GoogleOAuth2UserInfo(Map<String, Object> attributes) {
+  public GithubOAuth2UserInfo(Map<String, Object> attributes) {
     this.attributes = attributes;
   }
 
   @Override
   public String getProviderId() {
-    return (String) attributes.get("sub");
+    return String.valueOf(attributes.get("id"));
   }
 
   @Override
