+++ src/main/java/com/sequence/anonymous/security/userinfo/GoogleOAuth2UserInfo.java
@@ -1,18 +1,18 @@
-package com.sequence.anonymous.security;
+package com.sequence.anonymous.security.userinfo;
 
 import java.util.Map;
 
-public class GithubOAuth2UserInfo implements CustomOAuth2UserInfo {
+public class GoogleOAuth2UserInfo implements CustomOAuth2UserInfo {
 
   private final Map<String, Object> attributes;
 
-  public GithubOAuth2UserInfo(Map<String, Object> attributes) {
+  public GoogleOAuth2UserInfo(Map<String, Object> attributes) {
     this.attributes = attributes;
   }
 
   @Override
   public String getProviderId() {
-    return String.valueOf(attributes.get("id"));
+    return (String) attributes.get("sub");
   }
 
   @Override
